+++ lib/util.ts
@@ -1,21 +1,8 @@
 import { getCrypto } from '@/crypto';
 
-const importKey = async function (bytes: Uint8Array, name: string, usages: KeyUsage[]): Promise<CryptoKey> {
+const importKey = async function (bytes: Uint8Array, name: string, usages: KeyUsage[], extractable: boolean): Promise<CryptoKey> {
   const crypto = getCrypto();
-  return await crypto.subtle.importKey('raw', bytes, { name }, true, usages);
-};
-
-const mergeUint8Array = function (arrays: Uint8Array[]): Uint8Array {
-  const totalLength = arrays.reduce((sum, a) => sum + a.length, 0);
-  const result = new Uint8Array(totalLength);
-
-  let offset = 0;
-  for (const a of arrays) {
-    result.set(a, offset);
-    offset += a.length;
-  }
-
-  return result;
+  return await crypto.subtle.importKey('raw', bytes, { name }, extractable, usages);
 };
 
 const splitByLengths = function (array: Uint8Array, lengths: number[]): Uint8Array[] {
@@ -43,4 +30,4 @@ const splitByLengths = function (array: Uint8Array, lengths: number[]): Uint8Arr
   return result;
 };
 
-export { importKey, mergeUint8Array, splitByLengths };
+export { importKey, splitByLengths };